[C API] June 2019 update (#869)
* [c_api] use all relevant flags in compilation

* [c_api] Remove redundant IndexIVFFlat declarations

- From IndexIVF_c.h, already declared in IndexIVFFlat_c.h

* [c_api] type changes

- replace `long` with a more suitable type
- provide definitions for `faiss_component_t` and `faiss_distance_t`

* [c_api] Define CFLAGS and CUDACFLAGS

+++ c_api/AuxIndexStructures_c.cpp
@@ -89,7 +89,7 @@ int faiss_IDSelectorRange_new(FaissIDSelectorRange** p_sel, idx_t imin, idx_t im
 DEFINE_GETTER(IDSelectorBatch, int, nbits)
 DEFINE_GETTER(IDSelectorBatch, idx_t, mask)
 
-int faiss_IDSelectorBatch_new(FaissIDSelectorBatch** p_sel, long n, const idx_t* indices) {
+int faiss_IDSelectorBatch_new(FaissIDSelectorBatch** p_sel, size_t n, const idx_t* indices) {
     try {
         *p_sel = reinterpret_cast<FaissIDSelectorBatch*>(
             new IDSelectorBatch(n, indices)
